+++ lib/data/flow_button_type.dart
@@ -21,6 +21,7 @@ enum FlowButtonType with LocalizedEnum {
   String get localizationEnumName => "FlowButtonType";
 
   static const List<FlowButtonType> defaultOrder = [
+    FlowButtonType.eny,
     FlowButtonType.transfer,
     FlowButtonType.income,
     FlowButtonType.expense,