+++ lib/objectbox/objectbox.g.dart
@@ -455,7 +455,7 @@ final _entities = <obx_int.ModelEntity>[
   obx_int.ModelEntity(
     id: const obx_int.IdUid(10, 7829328581176695647),
     name: 'UserPreferences',
-    lastPropertyId: const obx_int.IdUid(24, 4516202653950496157),
+    lastPropertyId: const obx_int.IdUid(25, 5365639887585710264),
     flags: 0,
     properties: <obx_int.ModelProperty>[
       obx_int.ModelProperty(
@@ -589,6 +589,12 @@ final _entities = <obx_int.ModelEntity>[
         id: const obx_int.IdUid(24, 4516202653950496157),
         name: 'primaryAccountUuid',
         type: 9,
+        flags: 0,
+      ),
+      obx_int.ModelProperty(
+        id: const obx_int.IdUid(25, 5365639887585710264),
+        name: 'createTransactionsPerItemInScans',
+        type: 1,
         flags: 0,
       ),
     ],
@@ -1610,7 +1616,7 @@ obx_int.ModelDefinition getObjectBoxModel() {
         final primaryAccountUuidOffset = object.primaryAccountUuid == null
             ? null
             : fbb.writeString(object.primaryAccountUuid!);
-        fbb.startTable(25);
+        fbb.startTable(26);
         fbb.addInt64(0, object.id);
         fbb.addOffset(1, uuidOffset);
         fbb.addBool(2, object.combineTransfers);
@@ -1633,6 +1639,7 @@ obx_int.ModelDefinition getObjectBoxModel() {
         fbb.addBool(21, object.transactionListTileRelaxedDensity);
         fbb.addOffset(22, transactionEntryFlowJsonOffset);
         fbb.addOffset(23, primaryAccountUuidOffset);
+        fbb.addBool(24, object.createTransactionsPerItemInScans);
         fbb.finish(fbb.endTable());
         return object.id;
       },
@@ -1665,6 +1672,8 @@ obx_int.ModelDefinition getObjectBoxModel() {
             const fb.BoolReader().vTableGet(buffer, rootOffset, 26, false);
         final transactionListTileRelaxedDensityParam = const fb.BoolReader()
             .vTableGet(buffer, rootOffset, 46, false);
+        final createTransactionsPerItemInScansParam = const fb.BoolReader()
+            .vTableGet(buffer, rootOffset, 52, false);
         final trashBinRetentionDaysParam = const fb.Int64Reader()
             .vTableGetNullable(buffer, rootOffset, 14);
         final defaultFilterPresetParam = const fb.StringReader(
@@ -1719,6 +1728,8 @@ obx_int.ModelDefinition getObjectBoxModel() {
                     transactionListTileShowAccountForLeadingParam,
                 transactionListTileRelaxedDensity:
                     transactionListTileRelaxedDensityParam,
+                createTransactionsPerItemInScans:
+                    createTransactionsPerItemInScansParam,
                 trashBinRetentionDays: trashBinRetentionDaysParam,
                 defaultFilterPreset: defaultFilterPresetParam,
                 enableICloudSync: enableICloudSyncParam,
@@ -2574,6 +2585,10 @@ class UserPreferences_ {
   static final primaryAccountUuid = obx.QueryStringProperty<UserPreferences>(
     _entities[6].properties[21],
   );
+
+  /// See [UserPreferences.createTransactionsPerItemInScans].
+  static final createTransactionsPerItemInScans =
+      obx.QueryBooleanProperty<UserPreferences>(_entities[6].properties[22]);
 }
 
 /// [Budget] entity fields to define ObjectBox queries.